Break up into multiple files, and genericise the system
