+++ src/coreclr/tools/dotnet-pgo/MibcEmitter.cs
@@ -1,0 +1,337 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using Internal.TypeSystem;
+using Internal.TypeSystem.Ecma;
+using Internal.IL;
+using Microsoft.Diagnostics.Tracing;
+using Microsoft.Diagnostics.Tracing.Etlx;
+using Microsoft.Diagnostics.Tracing.Parsers.Clr;
+using System;
+using System.Buffers.Binary;
+using System.Collections.Generic;
+using System.IO;
+using System.Reflection;
+using System.Text;
+using System.Linq;
+using System.Diagnostics;
+using System.Globalization;
+using System.Threading.Tasks;
+using System.Reflection.Metadata;
+using System.Reflection.Metadata.Ecma335;
+using System.IO.Compression;
+using Microsoft.Diagnostics.Tracing.Parsers.Kernel;
+using System.Diagnostics.CodeAnalysis;
+using ILCompiler.Reflection.ReadyToRun;
+using Microsoft.Diagnostics.Tools.Pgo;
+using Internal.Pgo;
+
+namespace Microsoft.Diagnostics.Tools.Pgo
+{
+    interface IMibcEmitterTypeClass<MethodT, TypeT, TypeTOrUnknown>
+    {
+        EntityHandle GetMethodRef(TypeSystemMetadataEmitter emitter, MethodT method);
+        bool IsExactType(TypeTOrUnknown type);
+        TypeT GetExactType(TypeTOrUnknown type);
+        EntityHandle GetTypeRef(TypeSystemMetadataEmitter emitter, TypeT type);
+        int GetUnknownTypeIndex(TypeTOrUnknown type);
+        TypeT GetOwningType(MethodT method);
+        bool IsPrimitive(TypeT type);
+        bool IsCanonicalDefinitionType(TypeT type);
+        bool IsParameterizedType(TypeT type);
+        TypeT GetParameterType(TypeT type);
+        string GetAssemblyString(TypeT type);
+        ReadOnlySpan<TypeT> GetInstantiation(TypeT type);
+        ReadOnlySpan<TypeT> GetInstantiation(MethodT method);
+    }
+
+    static class MibcEmitter<MethodT, TypeT, TypeOrUnknownT, MibcEmitterTypeClass> where MibcEmitterTypeClass : IMibcEmitterTypeClass<MethodT, TypeT, TypeOrUnknownT>
+    {
+        class MIbcGroup : IPgoEncodedValueEmitter<TypeOrUnknownT>
+        {
+            private static int s_emitCount = 0;
+
+            public MIbcGroup(string name, TypeSystemMetadataEmitter emitter)
+            {
+                _buffer = new BlobBuilder();
+                _il = new InstructionEncoder(_buffer);
+                _name = name;
+                _emitter = emitter;
+            }
+
+            private BlobBuilder _buffer;
+            private InstructionEncoder _il;
+            private string _name;
+            private TypeSystemMetadataEmitter _emitter;
+
+            public void AddProcessedMethodData(ProcessedMethodData<MethodT> processedMethodData)
+            {
+                MethodT method = processedMethodData.Method;
+                string reason = processedMethodData.Reason;
+
+                // Format is 
+                // ldtoken method
+                // variable amount of extra metadata about the method, Extension data is encoded via ldstr "id"
+                // pop
+
+                // Extensions generated by this emitter:
+                //
+                // ldstr "ExclusiveWeight"
+                // Any ldc.i4 or ldc.r4 or ldc.r8 instruction to indicate the exclusive weight
+                //
+                // ldstr "WeightedCallData"
+                // ldc.i4 <Count of methods called>
+                // Repeat <Count of methods called times>
+                //  ldtoken <Method called from this method>
+                //  ldc.i4 <Weight associated with calling the <Method called from this method>>
+                //
+                // ldstr "InstrumentationDataStart"
+                // Encoded ints and longs, using ldc.i4, and ldc.i8 instructions as well as ldtoken <type> instructions
+                // ldstr "InstrumentationDataEnd" as a terminator
+                try
+                {
+                    EntityHandle methodHandle = default(MibcEmitterTypeClass).GetMethodRef(_emitter, method);
+                    _il.OpCode(ILOpCode.Ldtoken);
+                    _il.Token(methodHandle);
+                    if (processedMethodData.ExclusiveWeight != 0)
+                    {
+                        _il.LoadString(_emitter.GetUserStringHandle("ExclusiveWeight"));
+                        _il.LoadConstantI4(processedMethodData.ExclusiveWeight);
+                    }
+                    if (processedMethodData.WeightedCallData != null)
+                    {
+                        _il.LoadString(_emitter.GetUserStringHandle("WeightedCallData"));
+                        _il.LoadConstantI4(processedMethodData.WeightedCallData.Count);
+                        foreach (var entry in processedMethodData.WeightedCallData)
+                        {
+                            EntityHandle calledMethod = default(MibcEmitterTypeClass).GetMethodRef(_emitter, entry.Key);
+                            _il.OpCode(ILOpCode.Ldtoken);
+                            _il.Token(calledMethod);
+                            _il.LoadConstantI4(entry.Value);
+                        }
+                    }
+                    if (processedMethodData.InstrumentationData != null)
+                    {
+                        _il.LoadString(_emitter.GetUserStringHandle("InstrumentationDataStart"));
+                        PgoProcessor.EncodePgoData<TypeOrUnknownT>(processedMethodData.InstrumentationData, this, true);
+                    }
+                    _il.OpCode(ILOpCode.Pop);
+                }
+                catch (Exception ex)
+                {
+                    Program.PrintWarning($"Exception {ex} while attempting to generate method lists");
+                }
+            }
+
+            public MethodDefinitionHandle EmitMethod()
+            {
+                s_emitCount++;
+                string basicName = "Assemblies_" + _name;
+                if (_name.Length > 200)
+                    basicName = basicName.Substring(0, 200); // Cap length of name at 200, which is reasonably small.
+
+                string methodName = basicName + "_" + s_emitCount.ToString(CultureInfo.InvariantCulture);
+                return _emitter.AddGlobalMethod(methodName, _il, 8);
+            }
+
+            bool IPgoEncodedValueEmitter<TypeOrUnknownT>.EmitDone()
+            {
+                _il.LoadString(_emitter.GetUserStringHandle("InstrumentationDataEnd"));
+                return true;
+            }
+
+            void IPgoEncodedValueEmitter<TypeOrUnknownT>.EmitLong(long value, long previousValue)
+            {
+                if ((value <= int.MaxValue) && (value >= int.MinValue))
+                {
+                    _il.LoadConstantI4(checked((int)value));
+                }
+                else
+                {
+                    _il.LoadConstantI8(value);
+                }
+            }
+
+            void IPgoEncodedValueEmitter<TypeOrUnknownT>.EmitType(TypeOrUnknownT type, TypeOrUnknownT previousValue)
+            {
+                if (default(MibcEmitterTypeClass).IsExactType(type))
+                {
+                    _il.OpCode(ILOpCode.Ldtoken);
+                    _il.Token(default(MibcEmitterTypeClass).GetTypeRef(_emitter, default(MibcEmitterTypeClass).GetExactType(type)));
+                }
+                else
+                    _il.LoadConstantI4(default(MibcEmitterTypeClass).GetUnknownTypeIndex(type));
+            }
+
+        }
+
+        private static void AddAssembliesAssociatedWithType(TypeT type, HashSet<string> assemblies, out string definingAssembly)
+        {
+            definingAssembly = default(MibcEmitterTypeClass).GetAssemblyString(type);
+            assemblies.Add(definingAssembly);
+            AddAssembliesAssociatedWithType(type, assemblies);
+        }
+
+        private static void AddAssembliesAssociatedWithType(TypeT type, HashSet<string> assemblies)
+        {
+            if (default(MibcEmitterTypeClass).IsPrimitive(type))
+                return;
+
+            if (default(MibcEmitterTypeClass).IsCanonicalDefinitionType(type))
+                return;
+
+            if (default(MibcEmitterTypeClass).IsParameterizedType(type))
+            {
+                AddAssembliesAssociatedWithType(default(MibcEmitterTypeClass).GetParameterType(type), assemblies);
+            }
+            else
+            {
+                assemblies.Add(default(MibcEmitterTypeClass).GetAssemblyString(type));
+                foreach (var instantiationType in default(MibcEmitterTypeClass).GetInstantiation(type))
+                {
+                    AddAssembliesAssociatedWithType(instantiationType, assemblies);
+                }
+            }
+        }
+
+        private static void AddAssembliesAssociatedWithMethod(MethodT method, HashSet<string> assemblies, out string definingAssembly)
+        {
+            AddAssembliesAssociatedWithType(default(MibcEmitterTypeClass).GetOwningType(method), assemblies, out definingAssembly);
+            foreach (var instantiationType in default(MibcEmitterTypeClass).GetInstantiation(method))
+            {
+                AddAssembliesAssociatedWithType(instantiationType, assemblies);
+            }
+        }
+
+        public static int GenerateMibcFile(TraceTypeSystemContext tsc, FileInfo outputFileName, ICollection<ProcessedMethodData<MethodT>> methodsToAttemptToPlaceIntoProfileData, bool validate, bool uncompressed)
+        {
+            TypeSystemMetadataEmitter emitter = new TypeSystemMetadataEmitter(new AssemblyName(outputFileName.Name), tsc);
+
+            SortedDictionary<string, MIbcGroup> groups = new SortedDictionary<string, MIbcGroup>();
+            StringBuilder mibcGroupNameBuilder = new StringBuilder();
+            HashSet<string> assembliesAssociatedWithMethod = new HashSet<string>();
+
+            foreach (var entry in methodsToAttemptToPlaceIntoProfileData)
+            {
+                MethodT method = entry.Method;
+                assembliesAssociatedWithMethod.Clear();
+                AddAssembliesAssociatedWithMethod(method, assembliesAssociatedWithMethod, out string definingAssembly);
+
+                string[] assemblyNames = new string[assembliesAssociatedWithMethod.Count];
+                int i = 1;
+                assemblyNames[0] = definingAssembly;
+
+                foreach (string s in assembliesAssociatedWithMethod)
+                {
+                    if (s.Equals(definingAssembly))
+                        continue;
+                    assemblyNames[i++] = s;
+                }
+
+                // Always keep the defining assembly as the first name
+                Array.Sort(assemblyNames, 1, assemblyNames.Length - 1);
+                mibcGroupNameBuilder.Clear();
+                foreach (string s in assemblyNames)
+                {
+                    mibcGroupNameBuilder.Append(s);
+                    mibcGroupNameBuilder.Append(';');
+                }
+
+                string mibcGroupName = mibcGroupNameBuilder.ToString();
+                if (!groups.TryGetValue(mibcGroupName, out MIbcGroup mibcGroup))
+                {
+                    mibcGroup = new MIbcGroup(mibcGroupName, emitter);
+                    groups.Add(mibcGroupName, mibcGroup);
+                }
+                mibcGroup.AddProcessedMethodData(entry);
+            }
+
+            var buffer = new BlobBuilder();
+            var il = new InstructionEncoder(buffer);
+
+            foreach (var entry in groups)
+            {
+                il.LoadString(emitter.GetUserStringHandle(entry.Key));
+                il.OpCode(ILOpCode.Ldtoken);
+                il.Token(entry.Value.EmitMethod());
+                il.OpCode(ILOpCode.Pop);
+            }
+
+            emitter.AddGlobalMethod("AssemblyDictionary", il, 8);
+            MemoryStream peFile = new MemoryStream();
+            emitter.SerializeToStream(peFile);
+            peFile.Position = 0;
+
+            if (outputFileName.Exists)
+            {
+                outputFileName.Delete();
+            }
+
+            if (uncompressed)
+            {
+                using (FileStream file = new FileStream(outputFileName.FullName, FileMode.Create))
+                {
+                    peFile.CopyTo(file);
+                }
+            }
+            else
+            {
+                using (ZipArchive file = ZipFile.Open(outputFileName.FullName, ZipArchiveMode.Create))
+                {
+                    var entry = file.CreateEntry(outputFileName.Name + ".dll", CompressionLevel.Optimal);
+                    using (Stream archiveStream = entry.Open())
+                    {
+                        peFile.CopyTo(archiveStream);
+                    }
+                }
+            }
+
+            Program.PrintMessage($"Generated {outputFileName.FullName}");
+
+            if (validate)
+                return ValidateMIbcData(tsc, outputFileName, peFile.ToArray(), methodsToAttemptToPlaceIntoProfileData);
+            else
+                return 0;
+        }
+
+        static int ValidateMIbcData(TraceTypeSystemContext tsc, FileInfo outputFileName, byte[] moduleBytes, ICollection<ProcessedMethodData<MethodT>> methodsToAttemptToPrepare)
+        {
+            var mibcLoadedData = MibcReader.ReadMIbcData(tsc, outputFileName, moduleBytes).ToArray();
+            Dictionary<MethodT, MIbcData> mibcDict = new Dictionary<MethodT, MIbcData>();
+
+            foreach (var mibcData in mibcLoadedData)
+            {
+                mibcDict.Add((MethodT)mibcData.MetadataObject, mibcData);
+            }
+
+            bool failure = false;
+            if (methodsToAttemptToPrepare.Count != mibcLoadedData.Length)
+            {
+                Program.PrintError($"Not same count of methods {methodsToAttemptToPrepare.Count} != {mibcLoadedData.Length}");
+                failure = true;
+            }
+
+            foreach (var entry in methodsToAttemptToPrepare)
+            {
+                MethodT method = entry.Method;
+                if (!mibcDict.ContainsKey(method))
+                {
+                    Program.PrintError($"{method} not found in mibcEntryData");
+                    failure = true;
+                    continue;
+                }
+            }
+
+            if (failure)
+            {
+                return -1;
+            }
+            else
+            {
+                Program.PrintMessage($"Validated {outputFileName.FullName}");
+                return 0;
+            }
+        }
+
+    }
+}
